Removed unnecessary dependencies from Diorite Injections, now it only depends on asm/bytebuddy and annotation module.
